Upgrade to Eleventy 3.0
**Why:**

Eleventy 3.0 has been out as a canary (and beta) release for a while now
and it’s worked well on other, more complex projects, so we should be
good to upgrade here as well.

**What:**

Upgrade from Eleventy 2.0.0-canary.30 to 3.0.0-beta.1. Change Eleventy
config to use ESM.

Rename `workbox-config.js` to `workbox-config.cjs` to ensure it is
treated as CommonJS; update related npm script to look for new file
name.

Install latest version of `workbox-cli`.